model sync

+++ src/main/java/com/common/Player.java
@@ -1,6 +1,5 @@
 package com.common;
 import com.common.model.Decks.CommanderDeck;
-import com.common.model.Map.Map;
 import com.common.model.Map.MapNodes.MapNode;
 
 import java.util.ArrayList;
@@ -33,7 +32,7 @@ public class Player extends Message {
         return barrelSupply;
     }
 
-    public void setBarrelSupply(short barrelSupply) {
+    public void setBarrelSupply(int barrelSupply) {
         this.barrelSupply = barrelSupply;
     }
 
@@ -41,7 +40,7 @@ public class Player extends Message {
         return numberOfCastles;
     }
 
-    public void setNumberOfCastles(short numberOfCastles) {
+    public void setNumberOfCastles(int numberOfCastles) {
         this.numberOfCastles = numberOfCastles;
     }
 
@@ -49,7 +48,7 @@ public class Player extends Message {
         return ironThronePosition;
     }
 
-    public void setIronThronePosition(short ironThronePosition) {
+    public void setIronThronePosition(int ironThronePosition) {
         this.ironThronePosition = ironThronePosition;
     }
 
@@ -57,7 +56,7 @@ public class Player extends Message {
         return valyrianSwordPosition;
     }
 
-    public void setValyrianSwordPosition(short valyrianSwordPosition) {
+    public void setValyrianSwordPosition(int valyrianSwordPosition) {
         this.valyrianSwordPosition = valyrianSwordPosition;
     }
 
@@ -65,7 +64,7 @@ public class Player extends Message {
         return ravenPosition;
     }
 
-    public void setRavenPosition(short ravenPosition) {
+    public void setRavenPosition(int ravenPosition) {
         this.ravenPosition = ravenPosition;
     }
 
@@ -81,7 +80,7 @@ public class Player extends Message {
         return starNumber;
     }
 
-    public void setStarNumber(short starNumber) {
+    public void setStarNumber(int starNumber) {
         this.starNumber = starNumber;
     }
 
@@ -93,5 +92,4 @@ public class Player extends Message {
     @Override
     public void executeOnServer(Game game, ISocketManager socketManager) {
     }
-
 }
